Remove remaining RegionEndpoint usages
Remove specific RegionEndpoint field values from integration tests.
