+++ JustSaying.IntegrationTests/WhenRegisteringASqsSubscriber/WhenHandlingMultipleTopics.cs
@@ -1,11 +1,10 @@
 using System;
 using System.Threading.Tasks;
-using Amazon;
 using JustSaying.AwsTools.MessageHandling;
 using JustSaying.Messaging.MessageHandling;
 using JustSaying.Models;
 using JustSaying.TestingFramework;
-using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 using Newtonsoft.Json.Linq;
 using NSubstitute;
 using Xunit;
@@ -15,9 +14,6 @@ namespace JustSaying.IntegrationTests.WhenRegisteringASqsSubscriber
     [Collection(GlobalSetup.CollectionName)]
     public class WhenHandlingMultipleTopics : WhenRegisteringASqsTopicSubscriber
     {
-        public class TopicA : Message { }
-        public class TopicB : Message { }
-
         protected override Task When()
         {
             SystemUnderTest.WithSqsTopicSubscriber()
@@ -31,10 +27,21 @@ namespace JustSaying.IntegrationTests.WhenRegisteringASqsSubscriber
         [Fact]
         public async Task SqsPolicyWithAWildcardIsApplied()
         {
-            var queue = new SqsQueueByName(RegionEndpoint.EUWest1, QueueName, Client, 0, Substitute.For<ILoggerFactory>());
+            var queue = new SqsQueueByName(TestEnvironment.Region, QueueName, Client, 0, NullLoggerFactory.Instance);
+
             await Patiently.AssertThatAsync(() => queue.ExistsAsync(), TimeSpan.FromSeconds(60));
+
             dynamic policyJson = JObject.Parse(queue.Policy);
+
             policyJson.Statement.Count.ShouldBe(1,  $"Expecting 1 statement in Sqs policy but found {policyJson.Statement.Count}");
+        }
+
+        public class TopicA : Message
+        {
+        }
+
+        public class TopicB : Message
+        {
         }
     }
 }
